blogilistan laajennus step10

+++ osa4/controllers/blogs.js
@@ -16,17 +16,17 @@ blogsRouter.post('/', async (request, response, next) => {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
-  }
+    }
 
-  const user = await User.findById(decodedToken.id)
+    const user = await User.findById(decodedToken.id)
 
-  const blog = new Blog({
+    const blog = new Blog({
       title: request.body.title,
       author: request.body.author,
       url: request.body.url,
       likes: request.body.likes || 0,
       user: user._id
-  })
+    })
 
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
@@ -39,6 +39,20 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
+
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!request.token || !decodedToken.id) {
+          return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+        console.log(blog.user.toString())
+        console.log(user.id.toString())
+        if (blog.user.toString() !== user.id.toString()) {
+            return response.status(401).json({error: 'user not owner of this blog'})
+        }
+
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
       } catch (exception) {
